This commit fixes the integrative summary generation to preserve original formatting and make timecodes clickable, and resolves multiple `TypeError` and `ValueError` exceptions.
Key changes:
- The `VideoIntegrationThread` now appends new information as an HTML snippet to the existing summary, instead of regenerating it with an AI model. This preserves all original formatting.
- The `timestampDoubleClicked` signal is now connected for the `summaryTextArea`, enabling clickable timecodes that synchronize the video.
- Corrected a `ValueError` by properly parsing `MM:SS` timestamp strings into total seconds before calculation.
- Corrected a `TypeError` by ensuring the timestamp value passed to `QMediaPlayer.setPosition` is cast to an integer.
- Timecode formatting has been corrected to ensure reliable parsing by the `CustomTextEdit` widget.

+++ src/TGeniusAI.py
@@ -2285,7 +2285,7 @@ class VideoAudioManager(QMainWindow):
         Sincronizza il video al timestamp specificato in secondi.
         """
         if self.player:
-            self.player.setPosition(seconds * 1000)
+            self.player.setPosition(int(seconds * 1000))
 
     def setStartBookmark(self):
         self.videoSlider.setPendingBookmarkStart(self.player.position())
